:hammer: Code refactoring
