+++ services/blockchain-indexer/shared/dataService/business/interoperability/index.js
@@ -17,12 +17,8 @@ const { getBlockchainApps } = require('./blockchainApps');
 const { getChainAccount } = require('./chainAccount');
 const { getMainchainID } = require('./mainchain');
 const { getBlockchainAppsStatistics, reloadBlockchainAppsStats } = require('./blockchainAppsStats');
-const {
-	isMainchain,
-	resolveMainchainServiceURL,
-	resolveChannelInfo,
-	getCurrentChainID,
-} = require('./interoperability');
+const { isMainchain, getCurrentChainID } = require('./chain');
+const { resolveMainchainServiceURL, resolveChannelInfo } = require('./channel');
 
 module.exports = {
 	getBlockchainApps,
